chore: rename hook directory to plugin
