+++ engine/CMakeLists.txt
@@ -2,11 +2,11 @@ include(${CMAKE_CURRENT_SOURCE_DIR}/engine/config/CMakeLists.txt)
 include(${CMAKE_CURRENT_SOURCE_DIR}/engine/core/CMakeLists.txt)
 include(${CMAKE_CURRENT_SOURCE_DIR}/engine/entity/CMakeLists.txt)
 include(${CMAKE_CURRENT_SOURCE_DIR}/engine/event/CMakeLists.txt)
-include(${CMAKE_CURRENT_SOURCE_DIR}/engine/hook/CMakeLists.txt)
 include(${CMAKE_CURRENT_SOURCE_DIR}/engine/i18n/CMakeLists.txt)
 include(${CMAKE_CURRENT_SOURCE_DIR}/engine/input/CMakeLists.txt)
 include(${CMAKE_CURRENT_SOURCE_DIR}/engine/loader/CMakeLists.txt)
 include(${CMAKE_CURRENT_SOURCE_DIR}/engine/math/CMakeLists.txt)
+include(${CMAKE_CURRENT_SOURCE_DIR}/engine/plugin/CMakeLists.txt)
 include(${CMAKE_CURRENT_SOURCE_DIR}/engine/render/CMakeLists.txt)
 include(${CMAKE_CURRENT_SOURCE_DIR}/engine/resource/CMakeLists.txt)
 include(${CMAKE_CURRENT_SOURCE_DIR}/engine/script/CMakeLists.txt)
